endpoint/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mlflow==2.8.1
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
pandas==2.1.4
numpy==1.24.3
pytorch-lightning==2.1.3
torch==2.1.2
langgraph==0.0.20
openai==1.3.7
